requirements.txt
absl-py==2.3.1
accelerate==1.12.0
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
alembic==1.17.2
altair==5.5.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
appdirs==1.4.4
appnope==0.1.4
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asttokens==3.0.1
async-lru==2.0.5
asyncer==0.0.12
attrs==25.4.0
babel==2.17.0
backoff==2.2.1
bcrypt==5.0.0
beautifulsoup4==4.14.3
black==25.12.0
bleach==6.3.0
blinker==1.9.0
brotli==1.2.0
build==1.4.0
cachetools==5.5.2
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
chess==1.11.2
chromadb==1.4.1
click==8.3.1
cloudpickle==3.1.2
cmdstanpy==1.3.0
coloredlogs==15.0.1
colorlog==6.10.1
comm==0.2.3
contourpy==1.3.3
coreforecast==0.0.16
cryptography==46.0.3
curl_cffi==0.13.0
cycler==0.12.1
dataclasses-json==0.6.7
datasets==4.4.2
datasetsforecast==1.0.0
debugpy==1.8.19
decorator==5.2.1
defusedxml==0.7.1
dill==0.4.0
distro==1.9.0
durationpy==0.10
ecdsa==0.19.1
evaluate==0.4.6
executing==2.2.1
faiss-cpu==1.8.0
Farama-Notifications==0.0.4
fastapi==0.128.0
fastjsonschema==2.21.2
ffmpy==1.0.0
filelock==3.20.1
filetype==1.2.0
Flask==3.1.2
flatbuffers==25.12.19
fonttools==4.61.0
fqdn==1.5.1
frozendict==2.4.7
frozenlist==1.8.0
fsspec==2025.10.0
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.47.0
google-genai==1.59.0
googleapis-common-protos==1.72.0
gradio==6.3.0
gradio_client==2.0.3
groovy==0.1.2
grpcio==1.76.0
gymnasium==1.2.3
h11==0.16.0
hf-xet==1.2.0
holidays==0.87
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface-hub==0.36.0
humanfriendly==10.0
idna==3.11
ImageIO==2.37.2
importlib-metadata==6.11.0
importlib_resources==6.5.2
iniconfig==2.3.0
ipykernel==7.1.0
ipython==9.8.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.8
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.3
json5==0.13.0
json_repair==0.25.3
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.7.0
jupyter_core==5.9.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.5.1
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
jupyterlab_widgets==3.0.16
kaggle==1.8.3
kagglesdk==0.1.14
kiwisolver==1.4.9
kubernetes==35.0.0
langchain==0.1.14
langchain-community==0.0.38
langchain-core==0.1.53
langchain-text-splitters==0.0.2
langfuse==3.12.0
langsmith==0.1.81
lark==1.3.1
lazy_loader==0.4
librt==0.7.7
lightgbm==4.6.0
lightning-utilities==0.15.2
llvmlite==0.46.0
Mako==1.3.10
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
matplotlib==3.10.8
matplotlib-inline==0.2.1
mdurl==0.1.2
mistune==3.2.0
mmh3==5.2.0
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.0
multiprocess==0.70.18
multitasking==0.0.12
mypy==1.19.1
mypy_extensions==1.1.0
narwhals==2.13.0
nbclient==0.10.4
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.6.1
neuralforecast==3.1.2
nixtla==0.7.2
nixtlats==0.5.2
nltk==3.9.2
notebook==7.5.1
notebook_shim==0.2.4
numba==0.63.1
numpy==2.4.1
oauthlib==3.3.1
onnxruntime==1.23.2
openai==2.15.0
opencv-python==4.13.0.90
opencv-python-headless==4.11.0.86
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-grpc==1.39.1
opentelemetry-exporter-otlp-proto-http==1.39.1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
optuna==4.6.0
orjson==3.11.5
overrides==7.7.0
packaging==26.0
pandas==2.0.3
pandocfilters==1.5.1
parso==0.8.5
passlib==1.7.4
pathspec==0.12.1
patsy==1.0.2
pdf2image==1.17.0
peewee==3.18.3
pexpect==4.9.0
pillow==12.1.0
platformdirs==4.5.1
plotly==6.5.0
pluggy==1.6.0
pooch==1.8.2
posthog==5.4.0
prometheus_client==0.23.1
prompt_toolkit==3.0.52
propcache==0.4.1
prophet==1.2.1
protobuf==6.33.4
psutil==7.2.1
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.3
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.1.0
pydantic_core==2.41.5
pydeck==0.9.1
pydub==0.25.1
Pygments==2.19.2
PyJWT==2.10.1
PyMatting==1.1.14
pyparsing==3.2.5
pypdf==6.6.0
PyPDF2==3.0.1
PyPika==0.50.0
pyproject_hooks==1.2.0
pytesseract==0.3.13
pytest==9.0.2
python-chess==1.999
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-jose==3.5.0
python-json-logger==4.0.0
python-multipart==0.0.21
python-slugify==8.0.4
pytokens==0.3.0
pytorch-lightning==2.6.0
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
ray==2.53.0
redis==7.1.0
referencing==0.37.0
regex==2025.11.3
# Editable install with no version control (rembg==2.0.72.post2.dev0+00f918b)
-e /Users/dakshjain/Desktop/TestCases/Neo/NeoBRownfield/rembg
reportlab==4.4.7
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==13.9.4
rpds-py==0.30.0
rsa==4.9.1
safehttpx==0.1.7
safetensors==0.7.0
scikit-image==0.26.0
scikit-learn==1.8.0
scipy==1.16.3
seaborn==0.13.2
semantic-version==2.10.0
Send2Trash==2.0.0
sentence-transformers==2.6.0
sentencepiece==0.2.1
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8.1
SQLAlchemy==2.0.45
stack-data==0.6.3
stanio==0.5.1
starlette==0.50.0
statsmodels==0.14.6
streamlit==1.28.1
sympy==1.14.0
tenacity==8.5.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.4
terminado==0.18.1
text-unidecode==1.3
threadpoolctl==3.6.0
tifffile==2026.1.14
timm==1.0.24
tinycss2==1.4.0
tokenizers==0.22.1
toml==0.10.2
tomli==2.4.0
tomlkit==0.13.3
torch==2.9.1
torchmetrics==1.8.2
torchvision==0.24.1
tornado==6.5.3
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.3
typer==0.21.1
types-requests==2.32.4.20250913
types-tqdm==4.67.0.20250809
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uri-template==1.3.0
urllib3==2.6.2
utilsforecast==0.2.15
uv==0.9.17
uvicorn==0.40.0
uvloop==0.22.1
validators==0.35.0
watchdog==6.0.0
watchfiles==1.1.1
wcwidth==0.2.14
webcolors==25.10.0
webencodings==0.5.1
websocket-client==1.9.0
websockets==15.0.1
Werkzeug==3.1.4
wheel==0.45.1
widgetsnbextension==4.0.15
wrapt==1.17.3
xgboost==3.1.2
xlrd==2.0.2
xxhash==3.6.0
yarl==1.22.0
yfinance==1.0
zipp==3.23.0
zstandard==0.25.0
